Split packable impls into modules
